Revert: move the buffer traits back to binius-compute
Reverts #2253 at jimpo's request: "Please move the traits back to
binius-compute -- they make more sense in the context of the compute
memory/allocator modules."

BufferData and VecLike go back into crates/compute/src/lib.rs alongside
Allocator, together with the generic Vec and &mut [T] impls the orphan
rule forces to travel with them. crates/utils/src/buffer.rs and its
pub mod line are gone, and binius-compute drops the binius-utils
dependency #2253 added, so it is a leaf with no binius dependencies
again.

Two deliberate departures from a plain revert.

The pub use binius_compute::BufferData and its four-line comment in
field_buffer are not restored, per the same review: "If the pub use
reexport is a problem, just delete the re-export and comment." So the
26 call sites now name binius_compute directly rather than reaching for
the trait through binius_math::field_buffer, and there is one path to
say instead of three.

Two of the three unit tests #2253 added come along, covering
BufferData::truncate over Vec and over &mut [T]. Nothing in compute
exercised truncate before. Its third test, filling a Vec through
VecLike, is dropped as a duplicate of the existing
global_allocator_backs_a_plain_vec.
