+++ crates/compute/Cargo.toml
@@ -8,4 +8,3 @@ authors.workspace = true
 workspace = true
 
 [dependencies]
-binius-utils = { path = "../utils", default-features = false }
