+++ crates/hash/src/binary_merkle_tree.rs
@@ -3,10 +3,9 @@
 
 use std::{fmt, mem::MaybeUninit, slice};
 
-use binius_compute::{Allocator, GlobalAllocator};
+use binius_compute::{Allocator, GlobalAllocator, VecLike};
 use binius_field::Field;
 use binius_utils::{
-	buffer::VecLike,
 	checked_arithmetics::checked_log_2,
 	rayon::{self, prelude::*, slice::ParallelSlice},
 };
