+++ crates/ip-prover/src/prodcheck/mod.rs
@@ -2,16 +2,13 @@
 
 use std::iter;
 
-use binius_compute::Allocator;
+use binius_compute::{Allocator, VecLike};
 use binius_field::{Field, PackedField};
 use binius_ip::{mlecheck, prodcheck::MultilinearEvalClaim};
 use binius_math::{FieldBuffer, FieldVec, multilinear::hypercube::Hypercube};
-use binius_utils::{
-	buffer::VecLike,
-	rayon::{
-		prelude::*,
-		task_size::{IndexedParallelIteratorExt, WorkPerItem},
-	},
+use binius_utils::rayon::{
+	prelude::*,
+	task_size::{IndexedParallelIteratorExt, WorkPerItem},
 };
 use itertools::izip;
