+++ crates/math/src/multilinear/hypercube/expansion.rs
@@ -24,14 +24,11 @@
 
 use std::{iter, slice};
 
-use binius_compute::Allocator;
+use binius_compute::{Allocator, VecLike};
 use binius_field::{Field, PackedField, field::FieldOps};
-use binius_utils::{
-	buffer::VecLike,
-	rayon::{
-		prelude::*,
-		task_size::{IndexedParallelIteratorExt, WorkPerItem},
-	},
+use binius_utils::rayon::{
+	prelude::*,
+	task_size::{IndexedParallelIteratorExt, WorkPerItem},
 };
 
 use super::Hypercube;
