+++ crates/prover/src/protocols/shift/monster.rs
@@ -3,11 +3,10 @@
 
 use std::iter;
 
-use binius_compute::Allocator;
+use binius_compute::{Allocator, VecLike};
 use binius_core::{ShiftVariant, constraint_system::Shift, word::Word};
 use binius_field::{BinaryField, Field, PackedField};
 use binius_math::{FieldBuffer, FieldVec, multilinear::hypercube::Hypercube};
-use binius_utils::buffer::VecLike;
 use tracing::instrument;
 
 use super::{phase_1::SHIFT_OPERATOR_LOG_LEN, shift_ind::ShiftChallenge};
