+++ crates/utils/src/lib.rs
@@ -5,7 +5,6 @@
 //! Utility modules used in Binius.
 
 pub mod bitwise;
-pub mod buffer;
 pub mod checked_arithmetics;
 pub mod iter;
 #[cfg(feature = "platform-diagnostics")]
